Rename toolsController into CacheController
