+++ live/app/models/profile.js
@@ -1,0 +1,7 @@
+import DS from 'ember-data';
+const {Model, belongsTo, hasMany} = DS;
+
+export default Model.extend({
+  user: belongsTo('user'),
+  competences: hasMany('competence')
+});
