+++ live/mirage/factories/user.js
@@ -1,0 +1,3 @@
+import {Factory} from 'ember-cli-mirage';
+
+export default Factory.extend({});
